Update the Chain class with various improvements
I added a function to get all the data from all the blocks in the
current chain.

I added a function to return all the blocks in the chain.

I updated the verify chain function to throw an exception if it fails
and to return a boolean true if it does verify